Implement a minimal pallet for aleph
The pallet just stores a list of authorities. The implementation is inspired by the grandpa pallet, but uses macro attributes, like in the aura pallet.

A new module named primitives is added, and in that module the authority id type and the key type are defined, and the aleph runtime api is declared.