Make the parameters specifiable from command line arguments
It seems that the only place where we actually need to know the session_period and millisecs_per_block, is the command starting a node, and this knowledge is not neccessary in purge-chain and dev-keys.
This commit enables you to specify these parameters by adding "--session-period <period>" and/or "--millisecs-per-block <millis>".