A0-834: aleph-client: Always require std
In aleph-client, change the `std` feature requirements to be unconditional, and
remove the `std` feature (and the `[features]` section). This is accomplished
by opting back into default features, because a `std` feature is the only
default feature in all relevant crates.

Note that aleph-client already always required std (because it was not
`#[no_std]`).