Use good ole' run to install cargo-contract
It seems the baptiste0928/cargo-install action, while great at caching, is
not able to fix the version to *-beta.1 when a *-rc.1 version exists.
Using this method for now, as the rc line is not compatible with our
current node, it should be switched to the caching
baptiste0928/cargo-install action when a stable version is compatible
with the node.