Remove distinctions between `Full` and `Authority`
There were almost no nodes running without the `--validator` flag, and
the ones that were technically behaved wrongly due to a bug. This
removes the distinction, making all nodes run using the same code as
authority nodes. A hack was added, so that nodes running without
validator "pretend" to have an external address, but it's not used for
anything. If you weren't doing something wrong with your node before
this shouldn't impact you.