A0-2021: Use compact compressed runtime in workflows (#1406)
# Description

Currently, we produce compact wasm in the workflows, but we should use
compact.compressed to follow Polkadot approach, as well as to decrease
amount of bytes sent in the network during runtime upgrade.

This PR also removes not not-needed job which uploaded procution runtime
to S3. While this file was not overwritten, it is not good practice to
deploy runtime from non-push-to-main job.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

# Checklist:

<!-- delete when not applicable to your PR -->

- I have made neccessary updates to the Infrastructure