A0-4020: Pallet operations.  (#1637)
An account can have an underflow of a `consumers` counter.
Account categories that are impacted by this issue depend on a chain
runtime,
but specifically for AlephNode runtime are as follows:
* `consumers`  == 0, `reserved`  > 0
* `consumers` == 1, `balances.Locks` contain an entry with `id` ==
`vesting`
* `consumers` == 2, `balances.Locks` contain an entry with `id` ==
`staking`
* `consumers` == 3, `balances.Locks` contain entries with `id` ==
`staking`
   and account id is in `session.nextKeys`

This PR adds `pallet-operations` with a single extrinsic
`fix_accounts_consumers_underflow` that checks if the account falls into
one of the above categories, and increases its `consumers` counter in
such scenario.

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

There are 4 cases to test, and all of them are tested in the unit tests.
3 out of 4 cases were also tested on local chain with cloned state,
except 4th scenario (validator).

https://github.com/Cardinal-Cryptography/aleph-node/pull/1638

* weights adjustment for `fix_accounts_consumers_underflow`