L1-66: Pallet operations latest changes (#1758)
# Description

This PR extends and makes final logic changes for pallet operations.
Now, it works in general mode, which calculates the expected `consumers`
counter, compares it to the current one, and decreases or increases it
when necessary. There are two main additions to pallet logic:
* It is now aware of whether an account is a contract one - in such
cases it needs +1 `consumers`
* from `polkadot-sdk 1.4.0`, balance locks do not contribute to the
`consumers` counter anymore. Hence, that logic was removed from pallet
operations.

## Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

## Companion PRs

https://github.com/Cardinal-Cryptography/aleph-node/pull/1760