L1-290: Set the inflation parameters (#1820)
The Community has decided that the AZERO cap should be equal to 520M
AZERO, and the first year issuance equal to 27M AZERO.

The launch of the new inflation model is scheduled for October 14th. The
estimated total issuance at launch date is 373,948,365 AZERO.

Thus, the inflation horizon is set to 154,283,512,497 milliseconds
(approx. 1785 days).

- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

- I have added tests