L1-294: Safeguards for the inflation parameters setter (#1822)
We introduce safeguards that make it impossible to modify the values of
the inflation parameters in a way that the new value differs too much
from the old one. The rule is that the new value should lie between half
and twice the old value. With regard to the AZERO cap, we look at the
remaining issuance instead of the actual cap value, as this parameter
influences the payouts directly.

- Breaking change (fix or feature that would cause existing
functionality to not work as expected)