Don't request blocks you started iimporting
By optimistically assuming that a block that we started importing will
get imported, we treat it as already imported for purposes of sending
requests. This should limit both the amount of received blocks, as well
as the amount of blocks in the import queue, both of which were slowing
sync down.